torch.ao migration: quantize_jit.py phase1
Differential Revision: [D30880574](https://our.internmc.facebook.com/intern/diff/D30880574/)

[ghstack-poisoned]
